Add env variable for data source and change code for using this variable for choice parser service

+++ src/models/countrie.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String
+from sqlalchemy import Column, Integer, String, BigInteger
 
 from src.db.base import Base
 
@@ -9,7 +9,7 @@ class Country(Base):
     id = Column(Integer, primary_key=True, autoincrement=True)
     name = Column(String, nullable=False)
     region = Column(String, nullable=False)
-    population = Column(Integer, nullable=False)
+    population = Column(BigInteger, nullable=False)
 
     def __repr__(self):
         return f"<Country(name={self.name}, region={self.region}, population={self.population})>"
